refactor(ai-prompt-guard): migrate ai-prompt-guard to new framework
