+++ spec/03-plugins/42-ai-prompt-guard/02-integration_spec.lua
@@ -3,7 +3,7 @@ local helpers = require "spec.helpers"
 local PLUGIN_NAME = "ai-prompt-guard"
 
 
-for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
+for _, strategy in helpers.all_strategies() do
   describe(PLUGIN_NAME .. ": (access) [#" .. strategy .. "]", function()
     local client
 
@@ -511,4 +511,4 @@ for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
 
   end)
 
-end end
+end
